Bump fsevents to 1.2.11
fsevents does not provide binaries for node 12 and the package also
failed to build even if the dependencies for node-gyp appeared to be
installed.
Fortunately they released a new 1.2.11 version simplifying the module's
compilation.

This change only affects the development server under macOS, so it
should be safe to merge.